controlled components with control props
Sometimes it is desirable to allow a prop to be controlled via props,
instead of the component managing it externally. This means the user of
the component can set it manually. This is trivial if by doing that,
there is no need for the component to continue being able to manage it,
because it's done explicitly.

However, to support both use cases, whenever we would update that
particular bit of state, we need to check if we're receiving it as a
prop - and use that instead of the internal state. Otherwise, keep doing
as before.

In addition, whenever that state would be updated - for example, in
response to user input - then, instead of the component updating it's
own state, it needs to offer callbacks, which it calls instead, without
actually updating anything. This gives the user the opportunity to
update the state as they see fit.